+++ archive-product.php
@@ -1,1 +1,5 @@
-<?php woocommerce_content(); ?>
+<section class="page-section">
+  <div class="container">
+    <?php woocommerce_content(); ?>
+  </div>
+</section>