+++ lib/setup.php
@@ -13,11 +13,11 @@ function setup() {
   add_theme_support('soil-clean-up');
   add_theme_support('soil-disable-asset-versioning');
   add_theme_support('soil-disable-trackbacks');
-  add_theme_support('soil-nav-walker');
   add_theme_support('soil-nice-search');
+  add_theme_support('soil-relative-urls');
+  add_theme_support('soil-nav-walker');
   //add_theme_support('soil-js-to-footer');
   //add_theme_support('soil-jquery-cdn');
-  add_theme_support('soil-relative-urls');
 
   add_theme_support('woocommerce');
   add_theme_support( 'wc-product-gallery-zoom' );
@@ -66,8 +66,8 @@ function widgets_init() {
   register_sidebar([
     'name'          => __('Primary', 'cyon'),
     'id'            => 'sidebar-primary',
-    'before_widget' => '<section class="widget %1$s %2$s">',
-    'after_widget'  => '</section>',
+    'before_widget' => '<div class="widget col-sm-6 col-lg-12 %1$s %2$s">',
+    'after_widget'  => '</div>',
     'before_title'  => '<h4>',
     'after_title'   => '</h4>'
   ]);
@@ -75,8 +75,8 @@ function widgets_init() {
   register_sidebar([
     'name'          => __('Footer', 'cyon'),
     'id'            => 'sidebar-footer',
-    'before_widget' => '<section class="widget matchItem %1$s %2$s">',
-    'after_widget'  => '</section>',
+    'before_widget' => '<div class="widget col-sm-6 col-lg-3 %1$s %2$s">',
+    'after_widget'  => '</div>',
     'before_title'  => '<h4>',
     'after_title'   => '</h4>'
   ]);
@@ -84,8 +84,8 @@ function widgets_init() {
   register_sidebar([
     'name'          => __('Header Extra', 'cyon'),
     'id'            => 'sidebar-header',
-    'before_widget' => '<section class="widget %1$s %2$s">',
-    'after_widget'  => '</section>',
+    'before_widget' => '<div class="widget %1$s %2$s">',
+    'after_widget'  => '</div>',
     'before_title'  => '<h4>',
     'after_title'   => '</h4>'
   ]);