+++ templates/header.php
@@ -1,6 +1,6 @@
-<header class="banner banner-default hidden-md-down">
+<header class="banner banner-default d-none d-lg-block">
 	<?php if ( is_active_sidebar( 'sidebar-header' ) ) : ?>
-	<div class="secondary-nav">
+	<div class="secondary-nav bg-dark text-light">
 		<div class="container">
 			<?php dynamic_sidebar('sidebar-header'); ?>
 		</div>
@@ -8,12 +8,12 @@
 	<?php endif; ?>
 	<div class="primary-nav">
 		<div class="container">
-			<nav class="navbar navbar-toggleable-md">
+			<nav class="navbar navbar-expand-lg">
 				<a class="navbar-brand" href="<?= esc_url(home_url('/')); ?>"><?php bloginfo('name'); ?></a>
 				<div class="collapse navbar-collapse">
 		      <?php
 		      if (has_nav_menu('primary_navigation')) :
-		        wp_nav_menu(['theme_location' => 'primary_navigation', 'menu_class' => 'navbar-nav ml-auto']);
+		        wp_nav_menu(['theme_location' => 'primary_navigation', 'menu_class' => 'navbar-nav ml-auto', 'walker' => new WP_Bootstrap_Navwalker()]);
 		      endif;
 		      ?>
 				</div>
@@ -21,9 +21,9 @@
 		</div>
 	</div>
 </header>
-<header class="banner banner-mobile-default hidden-lg-up">
+<header class="banner banner-mobile-default d-lg-none">
 	<div class="container">
-		<nav class="navbar navbar-toggleable-md">
+		<nav class="navbar navbar-expand-lg">
 			<a class="navbar-brand" href="<?= esc_url(home_url('/')); ?>"><?php bloginfo('name'); ?></a>
 			<button class="navbar-toggler navbar-toggler-right menu-button" type="button">
 				<span class="sr-only">Main Navigation</span>