+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.2.0: October 14th, 2017
+* Updated to Bootstrap 4.0.0-beta
+* Updated to Sage 8.5.2
+* Removed most bower dependencies by default
+* Add Bootstrap nav walker
+
 ### 3.1.1: September 11th, 2017
 * Updated to Bootstrap 4.0.0-alpha.6
 * Use Flexnav for mobile nav